requirements.txt
# 机器学习依赖包
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# 深度学习框架
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0

# Transformer & 大语言模型
transformers==4.33.2
sentencepiece==0.1.99
accelerate==0.23.0

# 工具包
jupyter==1.0.0
ipykernel==6.25.1
tqdm==4.66.1
pyyaml==6.0.1
